decoupling secrets cli
